Updated SlotRealizers, pass Random via registry
The English language SlotRealizers are updated to reflect the data changes
in previous parts of the NewsEye PRA and the underlying data.

In addition, the SlotRealizers are now passed the Registry as a parameter,
rather than an instance of Random. The problem with Random was that it didn't
properly reset between subsequent calls, meaning that generating the 'same'
content multiple times produced varying results.

The current solution isn't perfect either, since it now resets TOO often,
not producing sufficient variety. Future work needs to be directed to
ensuring that the randomness is sufficiently, but not too, random.