Support for the compute_tf_idf utility
This utility computes absolute and relative counts as well as
TF-IDF values for various words. This commit includes a suitable
MessageGenerator as well as three templates, one each for each
type of value, in English. Other languages are at this point TODO.

Slot realization is done using the pre-existing [WORD:X] slot format.