Moved slot realizer registration to NewsPaperSlotRealizers
It was getting annoying to add new SlotRealizers to the system as it required
an explicit change to the NewspaperNlgService. I don't see a good reason why
we'd want to explicitly define all the realizers there. The only reason for
having this big list of realizers in the first place is to allow for disabled
SlotRealizers, but that can be achieved in this new implementation without
having to pollute the NlgService with over a hundred lines of code.