MINOR: Only import things present in __all__
numpy.random.RandomState is not, for some reason, mentioned in numpy.random's
__all__, so importing it directly is semantically a bit weird and raises some
automated warnings. This patch simply first imports numpy.random and then more
explicitly calls random.RandomState which sidesteps the warning but doesn't
change the weird semantics.

As RandomState is explicitly mentioned and described in the numpy docs as
a member of numpy.random, I'm not really sure why it's not in __all__.