Differentiate between no messages and no interesting messages
Previous implementation raised the same exception for failure to parse any messages at all
and all parsed messages having insufficient interestingness values for realization. This
distinction, however, is important for the user as the first case indicates are error in
the program while the second is an expected outcome.

This patch introduces a new NoInterestingMessagesException that is raised whenever there
are messages available for the generation task, but none of them are sufficiently
interesting to be included in the document.