Update numpy and transition to new RNG
The primary purpose of this patch is to update the version number of the
numpy dependency to 1.17. This version of numpy, however, uses a
significantly different style of RNG than the previous version used,
1.14.

While updating these changes, several instances of inconsistent use of
randomness were found. The patch also addresses those, resulting in a
more consistent use of randomness.

In the present state, python's random module is only used to determine
a shared seed that is used to reinitialize the actual prng every time
a new generation task is launched. Within the generation process, an
instance of numpy.random.Generator is used as the source of randomness.