REFACTOR: Move Matcher._equal_op out of the class
_equal_op is a function that defines the concept of equality within the
templating language. Originally a method of the Matcher class, it was
changed to a static method in commit f0505e0a. While that change made
the method's role more clear, it results in highly unpythonic code where
_equal_op.__func__ needs to be accessed to refer to the underlying
function within the Matcher class.

This patch simply moves the function out of Matcher, resulting in
cleaner and more pythonic code.