FIX: MessageGenerator defined as ABC twice
MessageGenerator was defined as an AbstractBaseClass, ABC, twice:
once in the definition of NLGPipelineComponent, it's parent class,
and once in its own definition. This resulted in the error

    TypeError: Cannot create a consistent method resolution
    order (MRO) for bases ABC, NLGPipelineComponent

This patch solves the issue by removing the ABC component from
MessageGenerator's class definition, but retaining it in that of
the parent class.