DOC: Better documentation for NLGPipelineComponent.run
The method NLGPipelineComponent suffers from a complication of
python's type hinting. This patch adjusts the function signature
slightly and provides the real type hint in a doc string, together
with a link to a github issue discussing the typing problem.

We'd want to define the function so that as long as the first
argument is of type Registry, we do not care what the types of the
latter arguments are. This, however, is not expressible using the
present type hinting system. For discussion, see e.g.
https://github.com/python/mypy/issues/5876