MAJOR: Per-processor templates
This is an initial step in a larger effort to separate the NLG resources
on a per-processor basis.

This commit introduces the concept of a `ProcessorResource` abstract
base class, creates multiple concrete implementations of such classes
and moves the phrase-level templates used by the system into these
resource classes.

NewspaperNLGService is also now aware of these ProcessorResources and
loads the templates from all the them to produce the total set of
templates used by the system.

Notably, while most of the ProcessorResources correspond to a specific
analytical tool, the NewspaperCorpusResource corresponds to the
underlying task as a whole, containing e.g. templates for realizing
information about the query underlying the analysis.

It's at this point clear that something along these lines is needed (as
the templates in NCRes do not map to any other ProcRes, but especially
the name is likely to change in the future.