MAJOR: Distributed MessageParsing, new TaskResult
This changeset distributes the process of parsing Message objects from
Investigator output to the various ProcessorResource classes.

Together with this change, the TaskResult class has been rewritten to
reflect changes into the PRA Investigator's output format.

NOTE: Current Investigator output is still lacking the fields related to
the query/dataset the result is about. These are expected to be added in
the immediate future. For now, the `corpus` and `corpus_type` fields are
simply filled with placeholder values.