Distribute SlotRealizerComponents
SlotRealizerComponents are now defined in a distributed manner in
various ProcessorResource classes. The logic behind this change is the
same as in the other distributive deltas: things that change together
(i.e. resources pertaining to a specific Investigator Processor) should
be located together.

In the future, we need to consider modifying the realization logic so
that when realizing the tokens of a result pertaining to Processor A, we
only check the SlotRealizerComponents related to either that specific
Processor and the SlotRealizerComponents related to the dataset being
described. This will likely be necessary to avoid any situations where
two Proc.Res. classes define SRCs which conflict.