CLEANUP: Major cleanup, autoformatter updates
After this changeset, the project uses a combination of isort, black and
flake8 for autoformatting and linting. This is part of an effort by the
undersigned to standardize on a specific formatting across multiple
projects.

All future commits should conform to whatever is dictated by running
 $ isort -rc . && black . && flake8 .

This patch also includes a pre-commit-hook to this effect, it's
suggested that anyone developing the software install the hook by
running
 $ pre-commit install