RegexRealizer slot requirements, attribute config
RegexRealizer now also takes a callable that defines a requirement for
the *slot* it's being applied to, `slot_requirements` in addition to
the previously available callable argument for defining requirements
for the *value*. This includes renaming `allowed` to
`group_requirements` to make the difference between the callables
clearer.

Furthermore, RegexRealizer's contructor now allows for an
`attach_attributes_to` list of indices. If provided, in a situations
where callign realize() produces multiple components, the attributes
(e.g. case) of the original attribute are attached only to the specified
new slots. Omitting this attribute attaches the attributes to all the
new slots. The change is thus backwards compatible.