Add simple system-wide test
The tests simply verify that, given the example inputs in
/test/resources, running the pipeline with various options does not
error out. The output content is not inspected beyond verifying a lack
of known basic error messages and exceptions.

The main purpose of these tests is to provide a simple way of ensuring
that the pipeline works as a whole in a manner that does not require
constant changes to the test to reflect minor changes in the system
output, cause by e.g. templating changes.