Improved and refactored document planning
The Document Planner has been rewritten and redesigned essentially from
scratch.

The new implementation supports, for example, iterative and gradual
selection of candidates based on the context of the previously selected
message, rather than on pure interestringness values. This means that
the process is more likely to group together more similar messages
within the paragraphs.

The changeset also refactors the domain-dependent parts of the document
planning process out of the core module, leaving the core (more)
domain-independent.