FIX: Too greedy aggregation
The aggregator used to aggregate sentences "There were X articles in
the French language" and "there were X articles published in L'ouvre"
into "There were X articles in the French language and published in the
L'ouvre".

This is problematic, since the natural inference for the reader is that
there is a set of articles, size X, s.t. all articles in X are both in
the French language and published in the newspaper L'ouvre. The
underlying data, however, is simply saying that there are two
(potentially completely distinct) sets of articles, both size X, s.t.
articles in one are in French and articles in the other were published
in the L'ouvre.

In other cases, e.g. "There were X articles in French and Finnish", it
is clear that there are two distinct sets of articles. But at the same
time, it's NOT clear whether both sets are of size X, or whether X is
the sum of the sets' sizes.

In any case, this kind of aggregation is a huge mess with extremely
difficult semantincs. This changeset solves the problem by simply
forbidding aggregation over the result_value slot which contains the
numeric information about the result. This means that the above examples
would get aggregated to:
1) There were X articles in the French language and X articles published
in the L'ouvre
2) There were X articles in French and X articles in Finnish.