FIX: Document planning issues
The document planner wasn't working quite as intended, with the
satellite selection sometimes including satellites that didn't fit the
theme of the paragraph in any way. These changes make the theming
restrictions a bit stricter.

As an upside, we should no longer see total theming violations. The
downside is that, I think, we might now see paragraphs with only a
nucleus. But that's probably acceptable.