Merge server error output
Server now outputs a single list of errors, rather than individual
errors for all kinds of things. This is needed for multi-dataset
generations to function: as they can result in an arbitrary number of
generation tasks, they need to be able to return an arbitrary number of
errors as well.