MAJOR: Support for multi-part generation
This changeset introduces support for multi-part generation tasks. The
system can now split a single input dataset into multiple based on the
search_query and dataset fields, running the pipeline for each such
split in turn.

The major change is that the server output's fields body and headline
are now potentially lists of strings. The interpretation here is that
the first headline goes with the first body.

Some backwards compability is provided by the fact that, in cases where
the output only contains a single body and head, they are returned
directly as string. This, however, could change based on feedback from
UI.