Implement heuristic for number rounding
Numbers are now rounded using a heuristic that should balance number
length and significant digits. The heuristic identifies the number of
digits needed s.t. the rounded number is strictly non-zero. That is, for
a number like '6.123', the identified rounding is '6'. For a number like
'0.0012345678' the identified rounding is '0.001'. The number is then
rounded to that amount and two additional decimals. '6.123' thus becomes
'6.12' and '0.001234' becomes' 0.00123'.

If the number is an integer, no rounding is applied and the value is
displayed as an integer without any decimals.