Add support for links in output
The /api/report/json end point now supports an extra field, "links",
which takes a boolean flag. If the field is omitted, it defaults to
false.

When "links" is true, the output text will contain link placeholders of
the form "[LINK:uuid]", where "uuid" is an analysis UUID that is
relevant to the preceding text and should be linked to. The construction
of the actual link is left to the caller of the API.

The non-json API does not, at the moment, support the "links" field.
Support is to be added once someone needs it.