FIX: Performance degrades over time
This changeset fixes an issue where every run of the generator appended
a new NumberRealizer to the resources during the SlotRealizer stage.
This caused the system to slow in performance over time. Effectively,
the slot realization is O(MN) where M is the number of realizers and N
is the length of the text. We were incrementing M for every run of the
software, which is quite catastrophic over time.