Language-aware message parsers
In some cases (most prominently ExtractNames), it's useful to be aware
of what language we are generating even when parsing messages. This
change adds generation target language as an argument to parse_messages,
thus making them aware of it. Most existing parsers -- and likely most
future parsers as well -- will just ignore this, but it's there for
those cases that benefit from it.