Language-specific entity names
In cases where SOLR can provide names in multiple languages (i.e. it can
link the entity), the generator now picks the report target language
name of the entity as the realization.

This is likely suboptimal in the following scenario: A Finnish-language
report if being produced about Swedish-language newspapers discussing
the location "Helsingfors". As the entity linker is able to identify
this location, the Finnish language report will discuss how the most
commonly discussed entity is "Helsinki", even if that string never
appears in the text.

While the above can be slightly confusing, the confusion is hopefully
limited by the fact that we are not quoting the location name. In
addition, it's notable that using the text-language entity name has the
same problem in the reverse direction while concurrently being
non-unique in situations where multilingual datasets are being analyzed.

While not perfect, this is the best way I can think of to do this for
now.