Support TrackNameSentiment tool
This is probably one of the most complex realizers so far, and kinda
stretches the limits of what is reasonably doable. We are "abusing" the
realizers to generate multi-sentence strings for each message.

This is, however, pretty much necessary to ensure that the various
factors are discussed in a reasonable order: while the min, max and mean
sections could be decomposed to separate re-orderable messages, we'd
need additional complex tooling to ensure that the first component
always contains the year range and the entity expression. While this
could probably be done with alternative templates and the use of context
in template selection, I expect the resulting text would be less fluent.

For now, this seems good enough, and can be revisited later if needed.