Better doc planning for multi-corpus docs
This change modifies the high-level document planning of multi-corpus
(i.e. multi-query) documents s.t. paragraphs about the same corpus are
grouped together and ordered by within-group-interestingness, and the
groups are then further ordered by group-maximal-interestingness. The
outputs are still limited to a maximal number of paragraphs, so the
resulting text shouldn't be too long.