FIX: Broken QueryTopicModelProcessor
Changes in output broke the assumption that there'd always be certain
metadata fields present in the "parameters" section of the input. This
update modifies the processor s.t. we can handle both the styles seen so
far, as well as a potential case where no metadata about the model is
available.