MINOR: Removed problematic comma from template
This comma is fine in isolation, but causes problems in aggregation. The
unaggregated sentence is acceptable even without the comma, and that
produces a better aggregated sentence, so I'm removing it for now.

A more fundamental solution would involve special casing commas somehow
in aggregation, and that seems like an un-optimal use of time at this
point.